backend/requirements.txt
# Core Django
Django==6.0.1

# Database
dj-database-url==2.2.0

# CORS
django-cors-headers==4.6.0

# Security and environment
python-decouple==3.8

# Production server (for Gunicorn)
gunicorn==23.0.0

# Whitenoise for static files
whitenoise==6.8.2

# Main Features
djangorestframework==3.15.2
django-ratelimit
requests
transformers

